move init logic to index page
