+++ src/Page/Login/index.js
@@ -1,8 +1,6 @@
-import React, { useState, useEffect } from "react";
-import { useDispatch } from "react-redux";
+import React, { useState } from "react";
 import { useNavigate } from "react-router-dom";
 import { useForm } from "react-hook-form";
-import { Preferences } from "@capacitor/preferences";
 import LoginBackground from "../../Assets/Login/background-login-page.png";
 import { connect } from "react-redux";
 import useAPI from "../../hooks/useAPI";
@@ -24,19 +22,6 @@ const LoginPage = props => {
   } = useForm();
   const navigate = useNavigate();
   const [api] = useAPI();
-
-  const dispatch = useDispatch();
-
-  useEffect(() => {
-    Preferences.get({ key: "token" })
-      .then(token => {
-        if (token.value) {
-          dispatch.user.setToken(token.value);
-          return navigate(`/home`);
-        }
-      })
-      .catch(err => console.err(err));
-  });
 
   const onSumbitLogin = async value => {
     if (value?.credential.length < 13) {
